Moved components to Home page & Fixed design bugs
