--- src/components/Contact-form/Contact-form.module.scss
@@ -1,61 +1,0 @@
-@import "styles/colors.scss";
-@import "styles/fonts.scss";
-
-.container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: space-evenly;
-    margin-top: 100px;
-    @include headline1;
-    background-color: $fill_blue;
-    color: $fill_white;
-    min-height:400px;
-    max-height: 100%;
-    position: relative;
-    width: 100%;
-    z-index: 1;
-}
-
-.title {
-    font-weight: 700;
-    font-size: 2.5rem;
-    text-align: center;
-}
-
-.text {
-    font-weight: 400;
-    font-size: 1.2rem;
-}
-
-.input{
-  background-color: $fill_blue;
-  color: $fill_white;
-  border: 0;
-  width: 360px;
-  height: 32px;
-  outline: none;
-  font-size: 1rem;
-}
-.input::placeholder{
-  color: #ffffffd2;
-}
-
-.input_border{
-  height: 1px;
-  width: 360px;
-  position:relative;
-  bottom: 10px;
-  background-color: $fill_white;
-}
-
-
-@media screen and (max-width: 425px){
-    .input{
-        width: 80%;
-
-    }
-    .input_border{
-        width: 80%;
-    }
-}
