--- src/pages/Home/Home-page.module.scss
@@ -1,5 +1,0 @@
-body{
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
