+++ src/pages/Home/components/Contact-form/Contact-form.module.scss
@@ -1,0 +1,78 @@
+@import "styles/colors.scss";
+@import "styles/fonts.scss";
+
+.wrapper {
+  height: 415px;
+}
+
+.container {
+  align-items: center;
+  background-color: $fill_blue;
+  box-sizing: border-box;
+  color: $fill_white;
+  display: flex;
+
+  flex-direction: column;
+
+  padding-top: 76px;
+  padding-bottom: 76px;
+  
+  width: 100%;
+  height: 415px;
+
+  left: 0;
+  position: absolute;
+  z-index: 1;
+}
+
+.title {
+  @include headline2;
+
+  margin-top: 0;
+  margin-bottom: 25px;
+}
+
+.text {
+  @include text_menu;
+
+  margin-top: 0;
+  margin-bottom: 60px;
+}
+
+.input{
+  @include text_form_btn;
+
+  background-color: transparent;
+  margin-bottom: 20px;
+
+  border: 0;
+  border-bottom: 1px solid $fill_white;
+  outline: none;
+
+  width: 361px;
+  height: 32px;
+}
+
+.input::placeholder{
+  color: #ffffffd2;
+}
+
+.input_border{
+  height: 1px;
+  width: 360px;
+  position:relative;
+  bottom: 10px;
+  background-color: $fill_white;
+}
+
+
+@media screen and (max-width: 425px){
+  .input{
+      width: 80%;
+
+  }
+
+  .input_border{
+      width: 80%;
+  }
+}
