+++ src/pages/Home/components/Representation-info/Representation-info.module.scss
@@ -2,55 +2,56 @@
 @import "styles/fonts.scss";
 
 .headerSection {
-  @include headline1;
-  width: 100%;
-  min-height: 630px;
-  max-height: 100%;
   background-color: $fill_blue;
+  box-sizing: border-box;
   color: $fill_white;
   display: flex;
+
+  padding-top: 155px;
+  padding-left: 151px;
+  padding-bottom: 184px;
+
+  width: 100%;
+  
   flex-direction: column;
   justify-content: center;
 }
 
-.section {
-  margin-left: 10%;
-  margin-top: 5%;
-  margin-right: 10%;
-}
-
 .slogan {
-  font-size: 3.5rem;
+  @include headline1;
+
+  margin-top: 0;
   margin-bottom: 40px;
 }
 
 .text {
+  @include text_description;
+
+  margin-top: 0;
   margin-bottom: 60px;
-  font-size:1.5rem;
-  font-weight: 400;
-  line-height: 130%;
 }
 
 .button {
-  width: 200px;
-  height: 60px;
-  margin-bottom: 10%;
-  color: $fill_blue;
+  @include text_button;
+
   background-color: $fill_white;
-  font-size: 18px;
-  line-height: 22px;
-  font-weight: 600;
   border: $fill_white;
   border-radius: 5px;
-  transition: linear all 0.2s;
+  color: $fill_blue;
   cursor: pointer;
-}
 
-.button:hover {
-  color: $fill_white;
-  background-color: #050505;
-  border: #050505;
+  width: 200px;
+  height: 60px;
+  
   transition: linear all 0.2s;
+
+  &:hover {
+    background-color: #050505;
+    border: #050505;
+    color: $fill_white;
+    
+    transition: linear all 0.2s;
+  }
 }
 
 @media screen and (max-width: 1024px) {
